More cleanup on Hillfort Presenter/Activity
